fix for issue #54 -- just moved YarnProgramImporterEditor class to its own .cs file, as the error suggested
