+++ Editor/YarnProgramImporterEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2a1a2642bf9259a4b95d13af4498fa9a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
